refactor(aztec-node): extract block and checkpoint reads into a provider module
Move getBlock, getBlockData, getBlocks, getCheckpoint, getCheckpoints and their private
checkpoint-context helpers into a new `NodeBlockProvider` collaborator under `src/modules/`, built over
the block source. The provider owns the response assembly (block_response_helpers) and the
block/checkpoint parameter normalization; `AztecNodeService` keeps thin delegating methods so the public
and RPC surface is unchanged.
